Add support for args like grep's --colour
An argument can now have an "equals default".  For example, the default value of --paginate can be 0; specifying --paginate on its own sets the value to 80; but the user can specify an explicit width with, say, --paginate=132.  Argument like this work in the same way as grep's --colour, where the default is none, a lone --colour is short for --colour=auto, but one can also specify --colour=always.

This patch also adapts to some deprecations and tightening up of @safety rules in DMD 2.072.